refactor: Ladder에서 LineRenderer에 대한 의존 제거
LineRenderer는 view패키지로 이동
